refactor: moves DebtSettleView
