+++ apps/debt/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from apps.debt import views
+from debt.views.debt_settle_view import DebtSettleView
 
 urlpatterns = [
-    path("settle/<int:pk>/", views.DebtSettleView.as_view(), name="transaction-settle"),
+    path("settle/<int:pk>/", DebtSettleView.as_view(), name="transaction-settle"),
 ]
